# swagger/swagger-server/src/main/resources/swagger-dev-openapi.yml
openapi: 3.0.2
info:
  title: Cult API v0
  version: 0.0.11
  x-olp-service:
    name: cult
    version: v0
  description: |-
    This API specification defines the interface for `Swagger Service`.
...
security:
   - sessionAuth: []
paths:
...
  /openapi:
    parameters:
      - $ref: '#/components/parameters/XCorrelationId'
      - $ref: '#/components/parameters/XRequestId'
    get:
      summary: Get openapi
      operationId: getOpenapi
      tags:
        - API Information
      description: |-
        Get the OpenApi specification of the service
      responses:
        '200':
          description: |-
            the openapi specification of the service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenapiSpecification'
            text/yaml:
              schema:
                $ref: '#/components/schemas/OpenapiSpecification'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/XCorrelationId'
            X-Request-ID:
              $ref: '#/components/headers/XRequestId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
        '5XX':
          $ref: '#/components/responses/ErrorResponse'
components:
  headers:
    XCorrelationId:
      description: |-
        Auto-generated ID, which uniquely identifies the request, available in the response.

        When contacting support with an inquiry regarding a specific request, provide the value of this header which will help troubleshooting the issue.

        If not present, one will be generated.
      schema:
        type: string
      example: f9bc3746-79e5-4ae5-ba87-30a6f3b9faa2
    XRequestId:
      description: |-
        User-provided token that can be used to trace a request or a group of requests sent to the service.

        If not present, one will be generated.
      schema:
        type: string
      example: REQ-7a0cdf45-c095-49ca-b1bd-6e12e153a5a7
  parameters:
...
    XCorrelationId:
      in: header
      name: X-Correlation-ID
      description: |-
        Auto-generated ID, which uniquely identifies the request, available in the response.

        When contacting support with an inquiry regarding a specific request, provide the value of this header which will help troubleshooting the issue.

        If not present, one will be generated.
      required: false
      schema:
        type: string
      example: f9bc3746-79e5-4ae5-ba87-30a6f3b9faa2
    XRequestId:
      in: header
      name: X-Request-ID
      description: |-
        User-provided token that can be used to trace a request or a group of requests sent to the service.

        If not present, one will be generated.
      required: false
      schema:
        type: string
      example: REQ-7a0cdf45-c095-49ca-b1bd-6e12e153a5a7
  schemas:
...
    ErrorResponse:
      title: ErrorResponse
      description: |-
        The error data for any specific failure response.

        This message format is used whenever the cult Service reports a failure to the caller.

        The error codes allocated to this API definition are in the range of `E311001` to `E311999`.
      type: object
      required:
        - title
        - status
        - code
        - cause
        - action
        - correlationId
      properties:
        title:
          description: |-
            Error title
          type: string
          example: Input data validation failed
        status:
          description: |-
            HTTP status code
          type: integer
          example: 4XX
        code:
          description: |-
            Error code.
          type: string
          example: E311001
        cause:
          description: |-
            The cause of the error
          type: string
          example: The input data in question does not meet validation rules
        action:
          description: |-
            Actionable instructions for the API consumer
          type: string
          example: Correct input data and retry request
        correlationId:
          $ref: '#/components/schemas/XCorrelationId'
        details:
          description: Collection of error details
          type: array
          items:
            $ref: '#/components/schemas/ErrorResponseDetailsElement'
    ErrorResponseDetailsElement:
      title: ErrorResponseDetailsElement
      description: Error response details element
      type: object
      properties:
        title:
          description: |-
            Error message title
          type: string
          example: Not accessible
        source:
          description: |-
            Reference to JSON path or Property Reference
          type: string
          example: resourceId
        message:
          description: |-
            Localized message string that describes the issue with the cited source.

            Corrective action for the caller should be provided if appropriate and known.
          type: string
          example: String is not accessible is not accessible
        messageTemplate:
          description: |-
            The template to display the message
          type: string
          example: validation.constraint.between
        messagePlaceholders:
          type: object
          additionalProperties: {}
          example:
            field: size
            min: 3
            max: 64
    ExternalErrorResponse:
      title: ExternalErrorResponse
      description: |-
        The error data for any specific failure response.

        This message format is used whenever the gateway reports an error
      type: object
      required:
        - error
        - error_description
      properties:
        error:
          description: Error title
          type: string
          example: Input data failed validation
        error_description:
          description: Descriptive error message.
          type: string
          example: Requested resource not found
...
    OpenapiSpecification:
      title: OpenapiSpecification
      description: API response for the openapi endpoint
      type: object
      additionalProperties: true
...
    XCorrelationId:
      title: XCorrelationId
      description: |-
        Auto-generated ID, which uniquely identifies the request, available in the response.

        When contacting support with an inquiry regarding a specific request, provide the value of this header which will help troubleshooting the issue.

        If not present, one will be generated.
      type: string
      example: f9bc3746-79e5-4ae5-ba87-30a6f3b9faa2
  responses:
...
    Unauthorized:
      description: The request requires user authentication
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExternalErrorResponse'
    Forbidden:
      description: The request cannot be fulfilled
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExternalErrorResponse'
...
    InternalServerError:
      description: The server encountered an unexpected condition that prevented it from fulfilling the request
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadGateway:
      description: The server received an invalid response from the upstream server
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: The server is not ready to handle the request
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GatewayTimeout:
      description: The server did not get a response in time
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      description: API response in case of errors.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/XCorrelationId'
        X-Request-ID:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID

#    Bearer:
#      type: http
#      scheme: bearer
#      bearerFormat: JWT
#      description: |-
#        A token obtained from a separate endpoint using client credentials and an OAuth 1.0a HMAC-SHA256 signed request.
#